+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Added SubmitAttestationsV2 endpoint.
 - Validator REST mode Electra block support
 - Added validator index label to `validator_statuses` metric
+- Added Validator REST mode use of Attestation V2 endpoints and Electra attestations
 
 ### Changed